+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### Added
 - Support for Container Image Signing
+- Support for Scheduled Compute Instance Suspension
 
 ## 4.20.0 (March 31, 2021)